+++ rlberry/agents/__init__.py
@@ -13,3 +13,6 @@ from .optql import OptQLAgent
 from .psrl import PSRLAgent
 from .rlsvi import RLSVIAgent
 from .ucbvi import UCBVIAgent
+
+# Importing tools
+from .stable_baselines import StableBaselinesAgent
