+++ rlberry/agents/stable_baselines/__init__.py
@@ -1,0 +1,1 @@
+from .stable_baselines import StableBaselinesAgent
